+++ lib/src/converters/timestamp.dart
@@ -29,3 +29,39 @@ class TimestampConverter implements JsonConverter<DateTime, num> {
     return object.millisecondsSinceEpoch / 1000;
   }
 }
+
+class DateConverter implements JsonConverter<DateTime, String> {
+  const DateConverter();
+
+  @override
+  DateTime fromJson(String json) {
+    return DateTime.parse(json);
+  }
+
+  @override
+  String toJson(DateTime object) {
+    return object.toDate();
+  }
+}
+
+class DateOrNullConverter implements JsonConverter<DateTime?, String?> {
+  const DateOrNullConverter();
+
+  @override
+  DateTime? fromJson(String? json) {
+    if (json == null) return null;
+    return DateTime.parse(json);
+  }
+
+  @override
+  String? toJson(DateTime? object) {
+    if (object == null) return null;
+    return object.toDate();
+  }
+}
+
+extension DateExtension on DateTime {
+  String toDate() {
+    return "$year-${month.toString().padLeft(2, '0')}-${day.toString().padLeft(2, '0')}";
+  }
+}